+++ app/Http/Requests/Film/SearchFilmsRequest.php
@@ -2,6 +2,11 @@
 
 namespace App\Http\Requests\Film;
 
+use App\Rules\Actor\ActorExistsRule;
+use App\Rules\Director\DirectorExistsRule;
+use App\Rules\Film\FilmProductionYearBetweenDatesRule;
+use App\Rules\Film\FilmTitleHasProperLengthRule;
+use App\Rules\Genre\GenreExistsRule;
 use Illuminate\Foundation\Http\FormRequest;
 
 /**
@@ -23,11 +28,14 @@ class SearchFilmsRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'title' => ['nullable', 'string', 'max:255'],
-            'prod_year' => ['nullable', 'integer', 'min:1800', 'max:3000'],
-            'actors' => ['nullable', 'array', 'exists:users,id'],
-            'directors' => ['nullable', 'array', 'exists:users,id'],
-            'genres' => ['nullable', 'array', 'exists:genres,id']
+            'title' => ['bail', 'nullable', 'string', new FilmTitleHasProperLengthRule()],
+            'prod_year' => ['bail', 'nullable', 'integer', new FilmProductionYearBetweenDatesRule()],
+            'actors' => ['nullable', 'array'],
+            'actors.*' => ['bail', 'required', 'integer', new ActorExistsRule()],
+            'directors' => ['nullable', 'array'],
+            'directors.*' => ['bail', 'required', 'integer', new DirectorExistsRule()],
+            'genres' => ['nullable', 'array'],
+            'genres.*' => ['bail', 'required', 'integer', new GenreExistsRule()]
         ];
     }
 }